### 2025-03-15 #7905 @editorialbot
[PRE REVIEW]: MultiVae: A Python package for Multimodal Variational Autoencoders on Partial Datasets.
**Submitting author:** @AgatheSenellart (<a href="http://orcid.org/0009-0000-3176-6461">Agathe Senellart</a>)
**Repository:** https://github.com/AgatheSenellart/MultiVae/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
