### 2025-03-24 2f44067
change the output shape of CVAE and adapt trainer to log images from CVAE as well
