+++ src/multivae/trainers/base/callbacks.py
@@ -9,6 +9,8 @@ import os
 import warnings
 import torch
 from typing import Literal
+from torchvision.utils import save_image
+import tempfile
 
 import numpy as np
 from pythae.models.base.base_config import BaseConfig
@@ -438,3 +440,127 @@ class WandbCallback(TrainingCallback):  # pragma: no cover
 
     def on_train_end(self, training_config: BaseTrainerConfig, **kwargs):
         self.run.finish()
+
+
+class MLFlowCallback(TrainingCallback):  # pragma: no cover
+    """
+    A :class:`TrainingCallback` integrating the experiment tracking tool
+    `mlflow` (https://mlflow.org/).
+
+    It allows users to store their configs, monitor their trainings
+    and compare runs through a graphic interface. To be able use this feature you will need:
+
+        - the package `mlfow` installed in your virtual env. If not you can install it with
+
+        .. code-block::
+
+            $ pip install mlflow
+    """
+
+    def __init__(self):
+        if not mlflow_is_available():
+            raise ModuleNotFoundError(
+                "`mlflow` package must be installed. Run `pip install mlflow`"
+            )
+
+        else:
+            import mlflow
+
+            self._mlflow = mlflow
+
+    def setup(
+        self,
+        training_config: BaseTrainerConfig,
+        model_config: BaseConfig = None,
+        project_name:str = None,
+        run_name: str = None,
+        **kwargs,
+    ):
+        """
+        Setup the MLflowCallback.
+
+        args:
+            training_config (BaseTrainerConfig): The training configuration used in the run.
+
+            model_config (BaseAEConfig): The model configuration used in the run.
+
+            run_name (str): The name to apply to the current run.
+        """
+        self.is_initialized = True
+
+        training_config_dict = training_config.to_dict()
+
+        if project_name is not None:
+            self._mlflow.set_experiment(project_name)
+
+        self._mlflow.start_run(run_name=run_name)
+
+        logger.info(
+            f"MLflow run started with run_id={self._mlflow.active_run().info.run_id}"
+        )
+        if model_config is not None:
+            model_config_dict = model_config.to_dict()
+
+            self._mlflow.log_params(
+                {
+                    **training_config_dict,
+                    **model_config_dict,
+                }
+            )
+
+        else:
+            self._mlflow.log_params({**training_config_dict})
+
+    def on_train_begin(self, training_config, **kwargs):
+        model_config = kwargs.pop("model_config", None)
+        if not self.is_initialized:
+            self.setup(training_config, model_config=model_config)
+
+    def on_log(self, training_config: BaseTrainerConfig, logs, **kwargs):
+        global_step = kwargs.pop("global_step", None)
+
+        logs = rename_logs(logs)
+        metrics = {}
+        for k, v in logs.items():
+            if isinstance(v, (int, float)):
+                metrics[k] = v
+
+        self._mlflow.log_metrics(metrics=metrics, step=global_step)
+
+    def on_train_end(self, training_config: BaseTrainerConfig, **kwargs):
+        self._mlflow.end_run()
+
+    def on_prediction_step(self, training_config: BaseTrainerConfig, **kwargs):
+        """Log all metrics or media contained in the metrics dictionary"""
+
+        global_step = kwargs.pop("global_step", None)
+
+        metrics_media = kwargs.pop("metrics_media", {})
+
+        images = metrics_media.pop('images',{})
+
+        # Save the reconstructions to folder and log to ml_flow
+        for key, image in images.items():
+            tmpdir = tempfile.gettempdir()
+            save_image(image, os.path.join(tmpdir, f'{key}.png'))
+            self._mlflow.log_artifact(os.path.join(tmpdir, f'{key}.png'),'images')
+
+        
[…]
