### 2025-04-02 #7905 @osorensen
:wave: @rudraprsd @obliviateandsurrender @rmorgan10 @powerfulbean @maruti-iitm would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
