### 2025-04-03 500af0a
add logginf dir in mlflow callback
