### 2025-04-16 1db1abf
add a function to save mlflow info at the end of training
