### 2025-04-26 AgatheSenellart/MultiVae#50 @powerfulbean
[JOSS] Error encountered when trying to run the quick start example locally
Hi,

Thank you for this great work, it contains a lot of benchmarks and models which are really helpful. However, when I tried to run the example code in a fresh python environment, I got the following error:

`
    model = MVTCAE(model_config)
  File "joss_review\packages\MultiVae\src\multivae\models\mvtcae\mvtcae_model.py", line 37, in __init__
    super().__init__(model_config, encoders, decoders)
  File "joss_review\packages\MultiVae\src\multivae\models\base\base_ae_model.py", line 63, in __init__
    raise AttributeError(
AttributeError: The provided number of input_dims 2 doesn'tmatch the number of modalities (None in model config
`

Details of the environment:

Python: 3.8.20
multivae: 1.0.0
The code I am using: [The Quick Start Code](https://github.com/AgatheSenellart/MultiVae?tab=readme-ov-file#quickstart)

corresponding JOSS thread: [](https://github.com/openjournals/joss-reviews/issues/7996#issuecomment-2773120469)

Thank you!

Best,
Jin
