### 2025-05-06 fd7f043
unify common configuration description accross models configuration
