### 2025-05-06 AgatheSenellart/MultiVae#50 @AgatheSenellart
Hi,

Thank you so much for pointing out this error. There was indeed an argument missing in the model configuration of the quickstart example. This argument used to be optional but no longer is and we forgot to change the example when we made this change. We are very sorry for this setback and we made the appropriate changes in a linked pull request: we changed the example and made the error message more explicit in case this argument is forgotten. Below is the modified example.
Please don't hesitate to reach out for any additional feedback !

```python
# Load a dataset
from multivae.data.datasets import MnistSvhn

train_set = MnistSvhn(data_path="./data", split="train", download=True)

# Instantiate your favorite model:
from multivae.models import MVTCAE, MVTCAEConfig

model_config = MVTCAEConfig(
    n_modalities=2, # ! the line that was missing !
    latent_dim=20,
    input_dims={"mnist": (1, 28, 28), "svhn": (3, 32, 32)},
)
model = MVTCAE(model_config)

# Define a trainer and train the model !
from multivae.trainers import BaseTrainer, BaseTrainerConfig

training_config = BaseTrainerConfig(learning_rate=1e-3, num_epochs=10)

trainer = BaseTrainer(
    model=model,
    train_dataset=train_set,
    training_config=training_config,
)
trainer.train()
```
