### 2025-05-06 1baee8f
1. Update quickstart example in the online documentation 2. improve robustness of the MVAE model with empty batches 3. Improve the error message raised when performing sanity check in the trainer
