+++ src/multivae/models/mvae/mvae_model.py
@@ -150,7 +150,7 @@ class MVAE(BaseMultiVAE):
 
         epoch = kwargs.pop("epoch", 1)
         # The annealing factor is updated each batch, so we need to know the idx of the batch in the epoch
-        batch_ratio = kwargs.pop("batch_ratio", 0) 
+        batch_ratio = kwargs.pop("batch_ratio", 0)
         if epoch >= self.warmup:
             beta = 1 * self.beta
         else:
@@ -195,10 +195,10 @@ class MVAE(BaseMultiVAE):
                 metrics["recon" + "_".join(sorted(s))] = subset_recon
             else:
                 subset_elbo = subset_kld = subset_recon = torch.tensor(
-                    0.0, requires_grad=True)
+                    0.0, requires_grad=True
+                )
                 len_batch = 0.0
             total_loss += subset_elbo
-            
 
         return ModelOutput(
             loss=total_loss, loss_sum=total_loss * len_batch, metrics=metrics
