+++ paper/paper.md
@@ -39,90 +39,95 @@ header-includes:
 
 # Summary
 
-In recent years, there has been a major boom in the development of multimodal
-machine learning models. Among open topics, representation (fusion) and generation of multimodal data are very active fields of research. Recently, Multimodal
+In recent years, multimodal machine learning has seen significant growth, especially in representation learning and data generation. Recently, Multimodal
 Variational Autoencoders (VAEs) have been attracting growing interest for both tasks, thanks
-to their versatility, scalability, and interpretability as probabilistic latent variable
-models. They are also particularly interesting in *partially observed*
-settings, as most can be trained even with missing data. 
-This last point makes them particularly suited for the medical field, where available datasets are often incomplete [@antelmi:2019; @aguila:poe].
+to their versatility, scalability, and interpretability as latent variable
+models. They are particularly useful in *partially observed*
+settings, such as medical applications, where available datasets are often incomplete [@antelmi:2019; @aguila:poe].
 
-We present
-MultiVae, an open-source Python library bringing together unified implementations of multimodal VAEs. It has been designed
-for easy and customizable use of these models on fully or partially observed data. This
-library also facilitates the development and benchmarking of new algorithms by integrating
-several benchmark datasets, a collection of evaluation metrics and tools for monitoring and
+We introduce
+MultiVae, an open-source Python library offering unified implementations of multimodal VAEs. It is designed
+for easy and customizable use of these models on fully or partially observed data. It facilitates the development and benchmarking of new algorithms by including
+standard benchmark datasets, evaluation metrics and tools for monitoring and
 sharing models. 
 
 ## Multimodal Variational Autoencoders
 
-Two main goals are commonly pursued in Multimodal Machine Learning:
+Multimodal VAEs aim to:
 (1) Learn a shared representation from multiple modalities;
-(2) Learn to generate one missing modality given the ones that are available.
+(2) Generate one missing modality from available ones.
 
-Multimodal Variational Autoencoders aim at solving both issues at the same time. These models learn a latent representation $z$ of all modalities in a lower dimensional common space and learn to *decode* $z$ to generate each modality.  
-Let $X = (x_1, x_2, ... x_M)$ contain $M$ modalities. In the VAE setting, we define an *encoder* distribution $q_{\phi}(z|X)$ projecting the observations to the latent space, and decoders distributions $(p_{\theta}(x_i|z))_{1 \leq i \leq M}$ translating the latent code $z$ back to the observations. Those distributions are parameterized by neural networks that are trained to minimize an objective function derived from variational inference. See [@kingma] to learn more about the VAE framework and [@suzuki_survey_2022] for a survey on multimodal VAEs. 
+These models learn a latent representation $z$ of all modalities in a lower dimensional space and learn to *decode* $z$ to generate each modality.  
[…]
