+++ paper/paper.md
@@ -60,16 +60,13 @@ Multimodal VAEs aim to:
 These models learn a latent representation $z$ of all modalities in a lower dimensional space and learn to *decode* $z$ to generate each modality.  
 Let $X = (x_1, x_2, ... x_M)$ contain $M$ modalities. In the VAE setting, we define an *encoder* distribution $q_{\phi}(z|X)$ projecting the observations to the latent space, and decoders distributions $(p_{\theta}(x_i|z))_{1 \leq i \leq M}$ translating the latent code $z$ back to observations. Those distributions are parameterized by neural networks that are trained to minimize an objective function derived from variational inference. See [@kingma] to learn more about the VAE framework and [@suzuki_survey_2022] for a survey on multimodal VAEs. 
 
-A key differentiator of multimodal VAEs relies in the choice of the encoder $q_{\phi}(z|X)$. They fall into three main categories:
-- *Aggregated models* [@wu:2018; @shi:2019; @sutter:2021] use a mean or product operation to combine modalities,
-- *Joint models* [@suzuki:2016; @vedantam:2018; @senellart:2023] use a neural network taking all modalities as input, 
-- *Coordinated models* [@wang_deep_2017; @tian:2019] use separate latent spaces with additional similarity constraints. 
+A key differentiator of multimodal VAEs relies in the choice of the encoder $q_{\phi}(z|X)$. They fall into three main categories: *Aggregated models* [@wu:2018; @shi:2019; @sutter:2021] use a mean or product operation to combine modalities, *Joint models* [@suzuki:2016; @vedantam:2018; @senellart:2023] use a neural network taking all modalities as input, and *Coordinated models* [@wang_deep_2017; @tian:2019] use separate latent spaces with additional similarity constraints. 
 
 ![Different types of multimodal VAEs \label{types_vae}](mvae_models_diagrams.png){width=100%}
 MultiVae unifies these approaches in a modular and extensible way. 
 
 Notably, aggregated models offer a natural way of *learning* on incomplete datasets: for an incomplete sample $X$, the encoding $z$ and the objective function can be computed using only available modalities.
-However, MultiVae is the first library to provide implementations of these models with built-in support for missing data, using masks during loss computation.
+MultiVae is the first library to provide implementations of these models with built-in support for missing data, using masks during loss computation.
 
 ## Data Augmentation
 Another application of VAEs is Data Augmentation (DA): by sampling new latent codes $z$ and decoding them, *fully synthetic multimodal* samples can be generated to augment a dataset. 
@@ -81,17 +78,13 @@ MultiVae includes a `multivae.samplers` module with several sampling strategies 
 Despite the usefulness of multimodal VAEs, the lack of easy-to-use and verified implementations might hinder 
 applicative research. MultiVae offers unified implementations, designed to be accessible even for non-specialists. We ensure reliability by reproducing key results from original papers whenever possible.
 
-Related libraries contain implementations of Multimodal VAEs: the [Multimodal VAE Comparison Toolkit](https://github.com/gabinsane/multimodal-vae-comparison) [@sejnova:2024], the [Pixyz](https://github.com/masa-su/pixyz/blob/main/examples/jmvae.ipynb)[@suzuki2023pixyz] library and `multi-view-ae` [@Aguila2023] that is most closely related and released while we were developing MultiVae.
+Related libraries contain implementations of Multimodal VAEs: the [Multimodal VAE Comparison Toolkit](https://github.com/gabinsane/multimodal-vae-comparison) [@sejnova:2024], [Pixyz]  (https://github.com/masa-su/pixyz/blob/main/examples/jmvae.ipynb)[@suzuki2023pixyz] and `multi-view-ae` [@Aguila2023] that is most closely related to us and released while we were developing MultiVae.
 
[…]
