+++ paper/paper.md
@@ -78,9 +78,9 @@ MultiVae includes a `multivae.samplers` module with several sampling strategies 
 Despite the usefulness of multimodal VAEs, the lack of easy-to-use and verified implementations might hinder 
 applicative research. MultiVae offers unified implementations, designed to be accessible even for non-specialists. We ensure reliability by reproducing key results from original papers whenever possible.
 
-Related libraries contain implementations of Multimodal VAEs: the [Multimodal VAE Comparison Toolkit](https://github.com/gabinsane/multimodal-vae-comparison) [@sejnova:2024], [Pixyz]  (https://github.com/masa-su/pixyz/blob/main/examples/jmvae.ipynb)[@suzuki2023pixyz] and `multi-view-ae` [@Aguila2023] that is most closely related to us and released while we were developing MultiVae.
+Related libraries contain implementations of Multimodal VAEs: the [Multimodal VAE Comparison Toolkit](https://github.com/gabinsane/multimodal-vae-comparison) [@sejnova:2024], [Pixyz](https://github.com/masa-su/pixyz/blob/main/examples/jmvae.ipynb)[@suzuki2023pixyz] and [multi-view-ae](https://github.com/alawryaguila/multi-view-AE) [@Aguila2023] that is most closely related to us and released while we were developing MultiVae.
 
- We compare in a summarizing table below, the different features of each work. MultiVae differs and complements existing software packages in key ways: for each model, we made sure to offer great flexibility on parameters' choices and to include all implementation details present in the original codes. It supports incomplete datasets, which we consider essential for real-life applications, as well as  generative samplers, **benchmark datasets** and **metrics** to facilitate research. 
+ We compare in a summarizing table below, the different features of each work. MultiVae differs and complements existing software packages in key ways: it supports **incomplete datasets**, which we consider essential for real-life applications, as well as **generative samplers**, **benchmark datasets** and **metrics** to facilitate research. It contains a large range of models with a great flexibility on parameters' choices and including all implementation details present in the original codes that improve performance.  
  
 
 ## List of Models and Features
