### 2025-05-07 #7996 @AgatheSenellart
Hi,
Thank you to all the reviewers for reviewing our paper and for the valuable feedback we have already received!
Following the editors' recommendation in the pre-review, we have slightly reduced the length of our paper and it now is below 1000 words without changing anything to the structure/ the references or semantic content.
Don't hesitate to reach out if you have any other issues or recommendations to improve the quality of our work.
