### 2025-05-19 #7996 @powerfulbean
I have completed my review of this software. The code is documented well and the paper's content is clear. I only have a minor fixing suggestion about the "Datasets" section in the ReadMe file - it seems a hyperlink is missing for 'here'. However, this is just a super minor issue, and I would like to recommend accepting this package for publication in JOSS. I also learnt a lot from it. Good work team!
