### 2025-05-19 #7996 @maruti-iitm
@osorensen forgot to check the box. I like the paper and the software. Below are some minor comments. I recommend accept.

- Please add a CITATION.cff
- Please add doc strings to the test folder; following the testing scripts code is a bit hard. There seems to be no doc strings there.
- Are doc strings formatted according to Google style? https://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_google.html
- Also, please use a standard code formatter such as `yapf`
