### 2025-05-20 #7996 @osorensen
Thanks @maruti-iitm. @AgatheSenellart, all the reviewers have completed their checklists now, but please address the points menioned by @maruti-iitm in [the post above](https://github.com/openjournals/joss-reviews/issues/7996#issuecomment-2892021622) as well as you can, and report back to me when done.
