### 2025-05-20 0bcbaaf
[fix] add Citation.cff
