+++ tests/test_mopoe.py
@@ -2,26 +2,25 @@ import os
 import shutil
 from copy import deepcopy
 
-import numpy as np
 import pytest
 import torch
-from encoders import Encoder_test, Encoder_test_multilatents
-from pythae.config import BaseConfig
+from .encoders import EncoderTest, EncoderTestMultilatents
 from pythae.models.base.base_utils import ModelOutput
 
 from multivae.data.datasets.base import IncompleteDataset, MultimodalBaseDataset
 from multivae.models.auto_model.auto_model import AutoModel
 from multivae.models.base.base_config import BaseAEConfig
 from multivae.models.mopoe import MoPoE, MoPoEConfig
-from multivae.models.nn.default_architectures import Decoder_AE_MLP, Encoder_VAE_MLP
+from multivae.models.nn.default_architectures import Decoder_AE_MLP
 from multivae.trainers.base.base_trainer import BaseTrainer
 from multivae.trainers.base.base_trainer_config import BaseTrainerConfig
 
 
-class Test_model:
+class TestMoPoE:
+    """Test class for MoPoE model."""
     @pytest.fixture(params=["complete", "incomplete"])
     def dataset(self, request):
-        # Create simple small dataset
+        """Create basic dataset"""
         data = dict(
             mod1=torch.randn((6, 2)),
             mod2=torch.randn((6, 3)),
@@ -43,6 +42,7 @@ class Test_model:
 
     @pytest.fixture(params=["one_latent_space", "multi_latent_spaces"])
     def archi_and_config(self, beta, request):
+        """Create model configurations for test"""
         if request.param == "one_latent_space":
             # Create an instance of mvae model
             config1 = BaseAEConfig(input_dim=(2,), latent_dim=5)
@@ -51,10 +51,10 @@ class Test_model:
             config4 = BaseAEConfig(input_dim=(4,), latent_dim=5)
 
             encoders = dict(
-                mod1=Encoder_test(config1),
-                mod2=Encoder_test(config2),
-                mod3=Encoder_test(config3),
-                mod4=Encoder_test(config4),
+                mod1=EncoderTest(config1),
+                mod2=EncoderTest(config2),
+                mod3=EncoderTest(config3),
+                mod4=EncoderTest(config4),
             )
 
             model_config = MoPoEConfig(
@@ -78,10 +78,10 @@ class Test_model:
             config4 = BaseAEConfig(input_dim=(4,), latent_dim=5, style_dim=3)
 
             encoders = dict(
-                mod1=Encoder_test_multilatents(config1),
-                mod2=Encoder_test_multilatents(config2),
-                mod3=Encoder_test_multilatents(config3),
-                mod4=Encoder_test_multilatents(config4),
+                mod1=EncoderTestMultilatents(config1),
+                mod2=EncoderTestMultilatents(config2),
+                mod3=EncoderTestMultilatents(config3),
+                mod4=EncoderTestMultilatents(config4),
             )
             model_config = MoPoEConfig(
                 n_modalities=4,
@@ -101,23 +101,27 @@ class Test_model:
 
     @pytest.fixture(params=[1.0, 1.5, 2.0])
     def beta(self, request):
+        """Test the model with different beta values"""
         beta = request.param
 
         return beta
 
     @pytest.fixture(params=[True, False])
     def custom(self, request):
+        """Test the model with or without custom architectures"""
         return request.param
 
     @pytest.fixture
     def model(self, archi_and_config, custom):
+        """Create the model for testing"""
         if custom:
             model = MoPoE(**archi_and_config)
         else:
             model = MoPoE(archi_and_config["model_config"])
         return model
 
-    def test_setup(self, model, custom, archi_and_config):
+    def test_init(self, model, custom, archi_and_config):
[…]
