+++ tests/test_jmvae_model.py
@@ -14,9 +14,11 @@ from multivae.models import JMVAE, AutoModel, JMVAEConfig
 from multivae.trainers import BaseTrainer, BaseTrainerConfig
 
 
-class Test_forward_and_predict:
+class TestJMVAE:
+    """Main class for testing the JMVAE model."""
     @pytest.fixture()
     def dataset(self):
+        """Dummy dataset. """
         data = dict(
             mod1=torch.Tensor([[1.0, 2.0], [4.0, 5.0]]),
             mod2=torch.Tensor([[67.1, 2.3, 3.0], [1.3, 2.0, 3.0]]),
@@ -41,6 +43,7 @@ class Test_forward_and_predict:
         ]
     )
     def config_and_architectures(self, request):
+        """Create configuration for JMVAE."""
         model_config = JMVAEConfig(n_modalities=3, **request.param[0])
 
         if request.param[1]:
@@ -70,37 +73,55 @@ class Test_forward_and_predict:
             decoders = None
         return dict(model_config=model_config, encoders=encoders, decoders=decoders)
 
-    def test_base_functions(self, dataset, config_and_architectures):
+    def test_setup(self, config_and_architectures):
+        """Test initialization of the JMVAE model. Check attributes. """
         model = JMVAE(**config_and_architectures)
 
         # test model setup
         assert model.alpha == config_and_architectures["model_config"].alpha
         assert model.beta == config_and_architectures["model_config"].beta
         assert model.latent_dim == config_and_architectures["model_config"].latent_dim
+    
 
-        # test forward
-        loss = model(dataset, epoch=2, warmup=2).loss
-        assert type(loss) == torch.Tensor
-        assert loss.size() == torch.Size([])
+    def test_forward(self, dataset, config_and_architectures):
+        """Test forward function. Check the output. """
 
-        # test encode and decode
+        model = JMVAE(**config_and_architectures)
+        output = model(dataset, epoch=2, warmup=2)
+        assert isinstance(output, ModelOutput)
+        assert isinstance(output.loss,torch.Tensor)
+        assert output.loss.size() == torch.Size([])
+
+    def test_encode_decode(self, dataset, config_and_architectures):
+        """Test the encode_function. Check that the output is 
+        an instance of ModelOutput and check the shape of the latent embeddings.
+        Check the decode function. """
+
+        model = JMVAE(**config_and_architectures)
         for return_mean in [True, False]:
+
+            ## Encode all modalities
+            # generate one latent sample conditioning on all modalities.
             outputs = model.encode(dataset, return_mean=return_mean)
-            assert outputs.one_latent_space
             embeddings = outputs.z
+
+            assert outputs.one_latent_space
             assert isinstance(outputs, ModelOutput)
             assert embeddings.shape == (2, model.latent_dim)
-
+            
+            # decode in modality mod1 and check the shape
             out_dec = model.decode(outputs, modalities="mod1")
             assert out_dec.mod1.shape == (2, 2)
 
+            # generate 2 latent samples and flatten the output
             outputs = model.encode(dataset, N=2, flatten=True, return_mean=return_mean)
             assert outputs.z.shape == (4, model.latent_dim)
 
             out_dec = model.decode(outputs, modalities="mod1")
             assert out_dec.mod1.shape == (4, 2)
 
-            # Encode conditioning on a subset of modalities
+            ## Encode one modality
+            # generate 1 sample
             embeddings = model.encode(
                 dataset, cond_mod=["mod1"], return_mean=return_mean
             )
[…]
