+++ tests/test_multimodal_dataset.py
@@ -1,4 +1,3 @@
-import os
 
 import numpy as np
 import pytest
@@ -7,9 +6,10 @@ from pythae.data.datasets import DatasetOutput
 from multivae.data.datasets.base import MultimodalBaseDataset
 
 
-class Test:
+class TestMultimodalDataset:
     @pytest.fixture
     def input_dataset_test(self):
+        """Create data for testing the MultimodalBaseDataset class"""
         data = dict(
             mod1=np.array([[1, 2], [4, 5]]),
             mod2=np.array([[67, 2, 3], [1, 2, 3]]),
@@ -18,6 +18,7 @@ class Test:
         return dict(data=data, labels=labels)
 
     def test_create_dataset(self, input_dataset_test):
+        """Test the init and __getitem__ function of MultimodalBaseDataset. """
         dataset = MultimodalBaseDataset(**input_dataset_test)
 
         sample = dataset[0]

[…]
