+++ src/multivae/models/jmvae/jmvae_model.py
@@ -64,7 +64,6 @@ class JMVAE(BaseJointModel):
     ):
         """Generate encodings conditioning on all modalities or a subset of modalities.
 
-
         Args:
             inputs (MultimodalBaseDataset): The data to encode.
             cond_mod (Union[list, str], optional): The modalities to use to compute the posterior
@@ -79,6 +78,7 @@ class JMVAE(BaseJointModel):
             AttributeError: _description_
 
                 Generate encodings conditioning on all modalities or a subset of modalities.
+
         Returns:
             ModelOutput instance with fields:
                 z (torch.Tensor (N, n_data, latent_dim))
@@ -194,10 +194,8 @@ class JMVAE(BaseJointModel):
     def _sample_from_poe_subset_exact(
         self, subset: list, data: dict, N=1, return_mean=False, flatten=False
     ):
+        """Sample from the product of experts for infering from a subset of modalities.
         """
-        Sample from the product of experts for infering from a subset of modalities.
-        """
-
         # Get all the experts' means and logvars
         mus, logvars = [], []
         for mod in subset:
