+++ src/multivae/models/joint_models/joint_model_config.py
@@ -5,8 +5,7 @@ from ..base.base_config import BaseMultiVAEConfig
 
 @dataclass
 class BaseJointModelConfig(BaseMultiVAEConfig):
-    """
-    This is the base config for joint models.
+    """This is the base config for joint models.
 
 
 
