+++ tests/test_nexus_model.py
@@ -18,6 +18,7 @@ from multivae.trainers import BaseTrainer, BaseTrainerConfig
 
 class TestNexus:
     """Class for testing the Nexus model"""
+
     @pytest.fixture(params=["complete", "incomplete"])
     def dataset(self, request):
         """Create a dummy dataset"""
@@ -180,7 +181,7 @@ class TestNexus:
         ]
     )
     def custom_config_archi(self, request):
-        """"Create a model configuration"""
+        """ "Create a model configuration"""
         encoders = dict()
         decoders = dict()
         top_encoders = dict()
@@ -236,7 +237,7 @@ class TestNexus:
         )
 
     def test_model_setup_with_custom_architectures(self, custom_config_archi):
-        """Test the model initialization with custom architectures. """
+        """Test the model initialization with custom architectures."""
         model = Nexus(**custom_config_archi)
 
         if custom_config_archi["model_config"].gammas is not None:
@@ -293,9 +294,9 @@ class TestNexus:
         assert model.joint_encoder == custom_config_archi["joint_encoder"]
 
     def test_model_setup_without_custom_architectures(self, custom_config_archi):
-        """Test the model initialization without custom architectures. 
-        In that case, default architectures should be used. """
-
+        """Test the model initialization without custom architectures.
+        In that case, default architectures should be used.
+        """
         model = Nexus(model_config=custom_config_archi["model_config"])
 
         if custom_config_archi["model_config"].gammas is not None:
@@ -340,9 +341,9 @@ class TestNexus:
         assert isinstance(model.joint_encoder, Encoder_VAE_MLP)
 
     def test_setup_with_wrong_attributes(self, custom_config_archi):
-        """Check that init raises errors when the attributes 
-        are not correct"""
-
+        """Check that init raises errors when the attributes
+        are not correct
+        """
         dict_config = custom_config_archi["model_config"].to_dict()
         dict_config.pop("name")
 
@@ -451,17 +452,19 @@ class TestNexus:
             return Nexus(model_config=custom_config_archi["model_config"])
 
     def test_forward(self, model, dataset):
-        """Test the forward method for Nexus. 
-        We check that the output is a ModelOutput with the loss tensor. """
+        """Test the forward method for Nexus.
+        We check that the output is a ModelOutput with the loss tensor.
+        """
         output = model(dataset, epoch=2)
         loss = output.loss
-        assert type(loss) == torch.Tensor
+        assert isinstance(loss, torch.Tensor)
         assert loss.size() == torch.Size([])
         assert loss.requires_grad
 
     def test_encode(self, model, dataset):
-        """Test the encode function for Nexus. 
-        We check the shape of the latent variables, depending on parameters. """
+        """Test the encode function for Nexus.
+        We check the shape of the latent variables, depending on parameters.
+        """
         for return_mean in [True, False]:
             # Encode one sample
             outputs = model.encode(dataset[0], return_mean=return_mean)
@@ -484,7 +487,7 @@ class TestNexus:
                 dataset, cond_mod="mod3", N=10, return_mean=return_mean
             ).z
             assert embeddings.shape == (10, 4, model.latent_dim)
-            
+
             # Encode the dataset conditioning on a subset of modalities
             embeddings = model.encode(
                 dataset, cond_mod=["mod2", "mod4"], return_mean=return_mean
@@ -492,16 +495,17 @@ class TestNexus:
             assert embeddings.shape == (4, model.latent_dim)
 
     def test_decode(self, model, dataset):
[…]
