+++ src/multivae/models/mvtcae/mvtcae_config.py
@@ -5,8 +5,7 @@ from ..base.base_config import BaseMultiVAEConfig
 
 @dataclass
 class MVTCAEConfig(BaseMultiVAEConfig):
-    """
-    This is the base config class for the MVTCAE model from
+    """This is the base config class for the MVTCAE model from
     'Multi-View Representation Learning via Total Correlation Objective' Neurips 2021.
     The code is based on the original implementation that can be found here :
     https://github.com/gr8joo/MVTCAE/blob/master/run_epochs.py
