+++ tests/test_mvae_model.py
@@ -16,6 +16,7 @@ from multivae.trainers import BaseTrainer, BaseTrainerConfig
 
 class TestMVAE:
     """Class for testing the MVAE model."""
+
     @pytest.fixture(params=["complete", "incomplete"])
     def dataset(self, request):
         """Create dummy dataset"""
@@ -38,7 +39,7 @@ class TestMVAE:
             dataset = IncompleteDataset(data=data, masks=masks, labels=labels)
 
         return dataset
-    
+
     @pytest.fixture
     def custom_architectures(self):
         """Create custom architectures for the MVAE model"""
@@ -79,7 +80,7 @@ class TestMVAE:
 
     @pytest.fixture(params=[True, False])
     def model(self, custom_architectures, model_config, request):
-        """Create the model for testing. """
+        """Create the model for testing."""
         custom = request.param
         if custom:
             model = MVAE(model_config, **custom_architectures)
@@ -88,7 +89,7 @@ class TestMVAE:
         return model
 
     def test_init(self, model, dataset, model_config):
-        """Check that the attributes are correctly set during init. """
+        """Check that the attributes are correctly set during init."""
         assert model.k == model_config.k
 
         expected_subsets = [
@@ -102,23 +103,23 @@ class TestMVAE:
         for s in expected_subsets:
             assert s in model.subsets
 
-    def test_forward(self, model ,dataset):
-        """Test the forward method of the model. We check that 
-        the output is a ModelOutput containing the loss. """
+    def test_forward(self, model, dataset):
+        """Test the forward method of the model. We check that
+        the output is a ModelOutput containing the loss.
+        """
         output = model(dataset, epoch=2)
         loss = output.loss
-        assert type(loss) == torch.Tensor
+        assert isinstance(loss, torch.Tensor)
         assert loss.size() == torch.Size([])
         assert loss.requires_grad
-    
-    def test_encode(self, model , dataset):
-        """Test the encode method of the model. 
-        We check that the generated samples have the right shape depending 
-        on the parameters. """
 
+    def test_encode(self, model, dataset):
+        """Test the encode method of the model.
+        We check that the generated samples have the right shape depending
+        on the parameters.
+        """
         # Try encoding and prediction
         for return_mean in [True, False]:
-            
             # Encode one datapoint
             outputs = model.encode(dataset[-1], return_mean=return_mean)
             assert outputs.one_latent_space
@@ -126,7 +127,7 @@ class TestMVAE:
             assert isinstance(outputs, ModelOutput)
             assert embeddings.shape == (1, 5)
 
-            # Encode one datapoint, generate two latent samples. 
+            # Encode one datapoint, generate two latent samples.
             embeddings = model.encode(dataset[-1], N=2, return_mean=return_mean).z
             assert embeddings.shape == (2, 1, 5)
 
@@ -148,9 +149,9 @@ class TestMVAE:
             assert embeddings.shape == (len(dataset), 5)
 
     def test_predict(self, model, dataset):
-        """Test the predict method of MVAE. 
-        We check the reconstructions shapes depending on the parameters. """
-
+        """Test the predict method of MVAE.
+        We check the reconstructions shapes depending on the parameters.
+        """
         # Reconstruct all modalities conditioning on mod2
         Y = model.predict(dataset, cond_mod="mod2")
         assert isinstance(Y, ModelOutput)
@@ -169,9 +170,8 @@ class TestMVAE:
         assert Y.mod1.shape == (len(dataset) * 10, 2)
         assert Y.mod2.shape == (len(dataset) * 10, 3)
 
-
[…]
