### 2025-05-21 b1159d3
[lint] ruff unsafe fixes
