+++ src/multivae/samplers/maf_sampler/maf_sampler.py
@@ -189,8 +189,7 @@ class MAFSampler(BaseSampler):
         return output
 
     def save(self, dir_path):
-        """Save the config and trained models
-        """
+        """Save the config and trained models."""
         super().save(dir_path=dir_path)
 
         if not self.is_fitted:
