+++ src/multivae/samplers/base/base_sampler_config.py
@@ -6,7 +6,6 @@ from pythae.config import BaseConfig
 
 @dataclass
 class BaseSamplerConfig(BaseConfig):
-    """BaseSampler config class.
-    """
+    """BaseSampler config class."""
 
     pass

[…]
