### 2025-05-21 64e65df
[clean] check docstrings + additional ruff linting.
