+++ src/multivae/models/joint_models/__init__.py
@@ -1,3 +1,4 @@
+"""Base class for models using a joint encoder."""
 from .joint_model import BaseJointModel
 from .joint_model_config import BaseJointModelConfig
 
