+++ src/multivae/__init__.py
@@ -1,0 +1,1 @@
+"""MultiVae package."""
