+++ src/multivae/samplers/__init__.py
@@ -1,3 +1,4 @@
+"""Samplers for unconditional generation."""
 from .base import BaseSampler, BaseSamplerConfig
 from .gaussian_mixture import GaussianMixtureSampler, GaussianMixtureSamplerConfig
 from .iaf_sampler import IAFSampler, IAFSamplerConfig
