+++ src/multivae/samplers/iaf_sampler/iaf_sampler_config.py
@@ -7,7 +7,7 @@ from ..base import BaseSamplerConfig
 class IAFSamplerConfig(BaseSamplerConfig):
     """This is the IAF sampler model configuration instance.
 
-    Parameters:
+    Args:
         n_made_blocks (int): The number of MADE model to consider in the IAF. Default: 2.
         n_hidden_in_made (int): The number of hidden layers in the MADE models. Default: 3.
         hidden_size (list): The number of unit in each hidder layer. The same number of units is
