+++ src/multivae/metrics/base/__init__.py
@@ -1,4 +1,5 @@
 """Base class for metrics modules."""
+
 from .evaluator_class import Evaluator
 from .evaluator_config import EvaluatorConfig
 
