+++ src/multivae/trainers/base/__init__.py
@@ -1,4 +1,5 @@
 """Base trainer for MultiVae's models."""
+
 from .base_trainer import BaseTrainer
 from .base_trainer_config import BaseTrainerConfig
 
