+++ src/multivae/samplers/__init__.py
@@ -1,4 +1,5 @@
 """Samplers for unconditional generation."""
+
 from .base import BaseSampler, BaseSamplerConfig
 from .gaussian_mixture import GaussianMixtureSampler, GaussianMixtureSamplerConfig
 from .iaf_sampler import IAFSampler, IAFSamplerConfig
