### 2025-05-21 769ae42
[fix] remove doublons project in pyproject.toml
