+++ pyproject.toml
@@ -1,7 +1,3 @@
-[project]
-name = "multivae"
-version = "1.0.0"
-
 [tool.isort]
 multi_line_output = 3
 include_trailing_comma = true
