+++ tests/test_mmvaePlus.py
@@ -5,17 +5,17 @@ from copy import deepcopy
 import numpy as np
 import pytest
 import torch
-from .encoders import EncoderTestMultilatents
-
 from pythae.models.base.base_utils import ModelOutput
 
 from multivae.data.datasets.base import IncompleteDataset, MultimodalBaseDataset
 from multivae.models.auto_model.auto_model import AutoModel
 from multivae.models.base.base_config import BaseAEConfig
 from multivae.models.mmvaePlus import MMVAEPlus, MMVAEPlusConfig
+from multivae.models.nn.default_architectures import Decoder_AE_MLP
 from multivae.trainers.base.base_trainer import BaseTrainer
 from multivae.trainers.base.base_trainer_config import BaseTrainerConfig
-from multivae.models.nn.default_architectures import Decoder_AE_MLP
+
+from .encoders import EncoderTestMultilatents
 
 
 class TestMMVAEPlus:
