+++ tests/test_jnf_model.py
@@ -353,8 +353,8 @@ class TestJNF:
             ]
         )
 
-        assert model_rec.encoders.cpu() is type(model.encoders.cpu())
-        assert model_rec.decoders.cpu() is type(model.decoders.cpu())
+        assert isinstance(model_rec.encoders.cpu(), type(model.encoders.cpu()))
+        assert isinstance(model_rec.decoders.cpu(), type(model.decoders.cpu()))
 
         optim_rec_state_dict = torch.load(os.path.join(checkpoint_dir, "optimizer.pt"))
 
@@ -460,8 +460,8 @@ class TestJNF:
             ]
         )
 
-        assert model_rec.encoders.cpu() is type(model.encoders.cpu())
-        assert model_rec.decoders.cpu() is type(model.decoders.cpu())
+        assert isinstance(model_rec.encoders.cpu(), type(model.encoders.cpu()))
+        assert isinstance(model_rec.decoders.cpu(), type(model.decoders.cpu()))
 
     def test_compute_nll(self, model, dataset):
         """Test the compute_nll function of the JNF model"""
