+++ tests/test_jmvae_model.py
@@ -347,8 +347,8 @@ class TestJMVAE:
             ]
         )
 
-        assert model_rec.encoders.cpu() is type(model.encoders.cpu())
-        assert model_rec.decoders.cpu() is type(model.decoders.cpu())
+        assert isinstance(model_rec.encoders.cpu(), type(model.encoders.cpu()))
+        assert isinstance(model_rec.decoders.cpu(), type(model.decoders.cpu()))
 
         optim_rec_state_dict = torch.load(os.path.join(checkpoint_dir, "optimizer.pt"))
 
@@ -456,8 +456,8 @@ class TestJMVAE:
             ]
         )
 
-        assert model_rec.encoders.cpu() is type(model.encoders.cpu())
-        assert model_rec.decoders.cpu() is type(model.decoders.cpu())
+        assert isinstance(model_rec.encoders.cpu(), type(model.encoders.cpu()))
+        assert isinstance(model_rec.decoders.cpu(), type(model.decoders.cpu()))
 
     def test_compute_nll(self, model, dataset):
         """Test the compute_nll function for the JMVAE model."""
