+++ tests/test_nexus_model.py
@@ -677,8 +677,8 @@ class TestNexus:
             ]
         )
 
-        assert model_rec.encoders.cpu() is type(model.encoders.cpu())
-        assert model_rec.decoders.cpu() is type(model.decoders.cpu())
+        assert isinstance(model_rec.encoders.cpu(), type(model.encoders.cpu()))
+        assert isinstance(model_rec.decoders.cpu(), type(model.decoders.cpu()))
 
         optim_rec_state_dict = torch.load(os.path.join(checkpoint_dir, "optimizer.pt"))
 
@@ -787,5 +787,5 @@ class TestNexus:
             ]
         )
 
-        assert model_rec.encoders.cpu() is type(model.encoders.cpu())
-        assert model_rec.decoders.cpu() is type(model.decoders.cpu())
+        assert isinstance(model_rec.encoders.cpu(), type(model.encoders.cpu()))
+        assert isinstance(model_rec.decoders.cpu(), type(model.decoders.cpu()))
