+++ tests/test_mhvae.py
@@ -518,8 +518,8 @@ class Test_MHVAE:
                 ]
             )
 
-            assert model_rec.encoders.cpu() is type(model.encoders.cpu())
-            assert model_rec.decoders.cpu() is type(model.decoders.cpu())
+            assert isinstance(model_rec.encoders.cpu(), type(model.encoders.cpu()))
+            assert isinstance(model_rec.decoders.cpu(), type(model.decoders.cpu()))
 
         optim_rec_state_dict = torch.load(os.path.join(checkpoint_dir, "optimizer.pt"))
 
@@ -631,5 +631,5 @@ class Test_MHVAE:
                 ]
             )
 
-            assert model_rec.encoders.cpu() is type(model.encoders.cpu())
-            assert model_rec.decoders.cpu() is type(model.decoders.cpu())
+            assert isinstance(model_rec.encoders.cpu(), type(model.encoders.cpu()))
+            assert isinstance(model_rec.decoders.cpu(), type(model.decoders.cpu()))
