+++ tests/test_trainer.py
@@ -564,7 +564,6 @@ class TestLogging:
         assert not os.path.exists(os.path.join(tmp_path, log_output_dir))
         trainer._get_file_logger(os.path.join(tmp_path, log_output_dir))
 
-
         assert os.path.exists(os.path.join(tmp_path, log_output_dir))
         assert os.path.exists(
             os.path.join(
