+++ .github/workflows/black.yml
@@ -1,0 +1,14 @@
+name: Lint
+
+on: [push, pull_request]
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: psf/black@stable
+      with:
+        options: "--check --verbose"
+        src: "./src"
+        use_pyproject: true
