+++ .github/workflows/black.yml
@@ -8,7 +8,4 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: psf/black@stable
-        with:
-            options: "--check --verbose"
-            src: "./src"
-            use_pyproject: true
+        
