### 2025-05-22 #7996 @AgatheSenellart
Thank you @powerfulbean, @maruti-iitm and @rudraprsd for your reviews !

@osorensen, we have addressed the points raised by @maruti-iitm with the following fixes in the last Pull Request:

- We have added a CITATION.cff file.
- We have added doc strings and comments in the tests to make them easier to follow.
- Our doc strings are indeed formatted according to [google style](https://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_google.html). Some doc strings were not properly formatted, so we reformatted them. We checked compliance to the google style format using [Ruff Linter](https://docs.astral.sh/ruff/faq/#does-ruff-support-numpy-or-google-style-docstrings).
- We use [Black](https://black.readthedocs.io/en/stable/#) as a code formatter. @maruti-iitm, did you notice specific parts of the code that didn't seem well formatted and that we should improve ? To ensure the published code remains well formatted, we added a github action that checks compliance to Black format on each push / pull request.
- We have fixed the failing link in our README.md mentioned in [this post](https://github.com/openjournals/joss-reviews/issues/7996#issuecomment-2889531226).

Let us know if there are any remaining issues we should address or other improvements that we could make!
