### 2025-05-23 AgatheSenellart/MultiVae#56 @osorensen
JOSS paper
I'll post my editorial comments to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/7996) as issues here. This is the first one.

Please indicate your full affiliation, including city and country.  "Université de Paris-Cité", "Inria" and "Inserm" are not sufficient.

Here are some examples in recent papers:

https://joss.theoj.org/papers/10.21105/joss.07608
https://joss.theoj.org/papers/10.21105/joss.05986
https://joss.theoj.org/papers/10.21105/joss.08070
