### 2025-05-23 AgatheSenellart/MultiVae#59 @osorensen
Add figure caption
This figure should have a number and a caption, and the figure should be referenced at least once in the main text.

![Image](https://github.com/user-attachments/assets/dbf21d89-03b5-4388-8ded-d4cf59477fbf)

https://github.com/openjournals/joss-reviews/issues/7996
