### 2025-05-23 d63dab1
[fix] affilation on one line with city, country and lab
