### 2025-05-23 b904e5b
[paper] add reference to Figure.
