### 2025-05-23 6927ec5
[paper] reference figure.
