### 2025-05-23 7bba98d
[paper] change way to reference
