### 2025-05-23 ce63540
[paper] change fig position
