### 2025-05-23 AgatheSenellart/MultiVae#59 @AgatheSenellart
Thank you for this comment. We have added a caption and a reference in the last PR.
