+++ paper/paper.bib
@@ -174,7 +174,7 @@ url = {https://doi.org/10.21105/joss.05093},
   volume = {8}, number = {85},
    pages = {5093}, 
    author = {Ana Lawry Aguila and Alejandra Jayme and Nina Montaña-Brown and Vincent Heuveline and Andre Altmann}, 
-   title = {Multi-view-AE: A Python package for multi-view autoencoder models}, journal = {Journal of Open Source Software}
+   title = {Multi-view-AE: A {P}ython package for multi-view autoencoder models}, journal = {Journal of Open Source Software}
 }
 
 @online{huggingface,
@@ -247,7 +247,7 @@ url = {https://doi.org/10.21105/joss.05093},
 }
 
 @misc{suzuki2023pixyz,
-      title={Pixyz: a Python library for developing deep generative models}, 
+      title={Pixyz: a {P}ython library for developing deep generative models}, 
       author={Masahiro Suzuki, Takaaki Kaneko and Yutaka Matsuo},
       journal = {Advanced Robotics},
       volume = {0},
@@ -328,7 +328,7 @@ url={https://openreview.net/forum?id=Rn8u4MYgeNJ}
   author       = {Thomas M. Sutter and
                   Imant Daunhawer and
                   Julia E. Vogt},
-  title        = {Multimodal Generative Learning Utilizing Jensen-Shannon-Divergence},
+  title        = {Multimodal Generative Learning Utilizing {J}ensen-{S}hannon-Divergence},
   journal      = {CoRR},
   volume       = {abs/2006.08242},
   year         = {2020},
@@ -342,7 +342,7 @@ url={https://openreview.net/forum?id=Rn8u4MYgeNJ}
 
 @article{aguila:poe,
 	author = {Lawry Aguila, Ana and Chapman, James and Altmann, Andre},
-	title = {Multi-modal Variational Autoencoders for&nbsp;Normative Modelling Across Multiple Imaging Modalities},
+	title = {Multi-modal Variational Autoencoders for Normative Modelling Across Multiple Imaging Modalities},
 	year = {2023},
 	isbn = {978-3-031-43906-3},
 	publisher = {Springer-Verlag},
@@ -351,8 +351,6 @@ url={https://openreview.net/forum?id=Rn8u4MYgeNJ}
 	doi = {10.1007/978-3-031-43907-0_41},
 	abstract = {One of the challenges of studying common neurological disorders is disease heterogeneity including differences in causes, neuroimaging characteristics, comorbidities, or genetic variation. Normative modelling has become a popular method for studying such cohorts where the ‘normal’ behaviour of a physiological system is modelled and can be used at subject level to detect deviations relating to disease pathology. For many heterogeneous diseases, we expect to observe abnormalities across a range of neuroimaging and biological variables. However, thus far, normative models have largely been developed for studying a single imaging modality. We aim to develop a multi-modal normative modelling framework where abnormality is aggregated across variables of multiple modalities and is better able to detect deviations than uni-modal baselines. We propose two multi-modal VAE normative models to detect subject level deviations across T1 and DTI data. Our proposed models were better able to detect diseased individuals, capture disease severity, and correlate with patient cognition than baseline approaches. We also propose a multivariate latent deviation metric, measuring deviations from the joint latent space, which outperformed feature-based metrics.},
 	booktitle = {Medical Image Computing and Computer Assisted Intervention – MICCAI 2023: 26th International Conference, Vancouver, BC, Canada, October 8–12, 2023, Proceedings, Part I},
-	pages = {425–434},
-	numpages = {10},
 	keywords = {Unsupervised learning, Normative modelling, Multimodal modelling, multi-view VAEs},
 	location = {Vancouver, BC, Canada}
 }
