### 2025-06-05 260565e
reverte delete pages
