+++ paper/paper.bib
@@ -351,6 +351,8 @@ url={https://openreview.net/forum?id=Rn8u4MYgeNJ}
 	doi = {10.1007/978-3-031-43907-0_41},
 	abstract = {One of the challenges of studying common neurological disorders is disease heterogeneity including differences in causes, neuroimaging characteristics, comorbidities, or genetic variation. Normative modelling has become a popular method for studying such cohorts where the ‘normal’ behaviour of a physiological system is modelled and can be used at subject level to detect deviations relating to disease pathology. For many heterogeneous diseases, we expect to observe abnormalities across a range of neuroimaging and biological variables. However, thus far, normative models have largely been developed for studying a single imaging modality. We aim to develop a multi-modal normative modelling framework where abnormality is aggregated across variables of multiple modalities and is better able to detect deviations than uni-modal baselines. We propose two multi-modal VAE normative models to detect subject level deviations across T1 and DTI data. Our proposed models were better able to detect diseased individuals, capture disease severity, and correlate with patient cognition than baseline approaches. We also propose a multivariate latent deviation metric, measuring deviations from the joint latent space, which outperformed feature-based metrics.},
 	booktitle = {Medical Image Computing and Computer Assisted Intervention – MICCAI 2023: 26th International Conference, Vancouver, BC, Canada, October 8–12, 2023, Proceedings, Part I},
+	pages = {425–434},
+	numpages = {10},
 	keywords = {Unsupervised learning, Normative modelling, Multimodal modelling, multi-view VAEs},
 	location = {Vancouver, BC, Canada}
 }
