### 2025-06-05 #7996 @AgatheSenellart
Hi @crvernon,
I have updated the bib file to address the mentioned issues.
Let me know if there is anything more I can do.
